equatiomatic: Transform Models into 'LaTeX' Equations

The goal of 'equatiomatic' is to reduce the pain associated with writing 'LaTeX' formulas from fitted models. The primary function of the package, extract_eq(), takes a fitted model object as its input and returns the corresponding 'LaTeX' code for the model.

See also

Author

Maintainer: Philippe Grosjean phgrosjean@sciviews.org (ORCID)

Authors:

Daniel Anderson daniela@uoregon.edu (ORCID)

Andrew Heiss andrew@andrewheiss.com (ORCID)

Jay Sumners jay.sumners@gmail.com

Other contributors:

Joshua Rosenberg jmrosenberg@utk.edu (ORCID) [contributor]

Jonathan Sidi yonis@metrumrg.com (ORCID) [contributor]

Ellis Hughes ellishughes@live.com (ORCID) [contributor]

Thomas Fung thomas.fung.dr@gmail.com (ORCID) [contributor]

Reza Norouzian rnorouzian@gmail.com (ORCID) [contributor]

Indrajeet Patil patilindrajeet.science@gmail.com (ORCID) [contributor]

Quinn White quinnarlise@gmail.com (ORCID) [contributor]

David Kane dave.kane@gmail.com [contributor]

Teun van den Brand t.vandenbrand@nki.nl (ORCID) [contributor]

Guyliann Engels guyliann.engels@umons.ac.be (ORCID) [contributor]