LaTeX code for R models

Extract the variable names from a model to produce a LaTeX equation, which is output to the screen. Supports any model supported by broom::tidy.

extract_eq(model, ital_vars = FALSE, wrap = FALSE,
  terms_per_line = 4, operator_location = "end",
  align_env = "aligned", use_coefs = FALSE, coef_digits = 2,
  fix_signs = TRUE)

Arguments

model
A fitted model

ital_vars
Logical, defaults to FALSE. Should the variable names not be wrapped in the \text{} command?

wrap
Logical, defaults to FALSE. Should the terms on the right-hand side of the equation be split into multiple lines? This is helpful with models with many terms.

terms_per_line
Integer, defaults to 4. The number of right-hand side terms to include per line. Used only when wrap is TRUE.

operator_location
Character, one of “end” (the default) or “start”. When terms are split across multiple lines, they are split at mathematical operators like +. If set to “end”, each line will end with a trailing operator (+ or -). If set to “start”, each line will begin with an operator.

align_env
TeX environment to wrap around equation. Must be one of aligned, aligned*, align, or align*. Defaults to aligned.

use_coefs
Logical, defaults to FALSE. Should the actual model estimates be included in the equation instead of math symbols?

coef_digits
Integer, defaults to 2. The number of decimal places to round to when displaying model estimates.

fix_signs
Logical, defaults to FALSE. If disabled, coefficient estimates that are negative are preceded with a "+" (e.g. 5(x) + -3(z)). If enabled, the "+ -" is replaced with a "-" (e.g. 5(x) - 3(z)).

Value

A character of class “equation”.

Examples

# Simple model
mod1 <- lm(mpg ~ cyl + disp, mtcars)
extract_eq(mod1)

#> $$
#> \text{mpg} = \alpha + \beta_{1}(\text{cyl}) + \beta_{2}(\text{disp}) + \epsilon
#> $$

# Include all variables
mod2 <- lm(mpg ~ ., mtcars)
extract_eq(mod2)

#> $$
#> \text{mpg} = \alpha + \beta_{1}(\text{cyl}) + \beta_{2}(\text{disp}) + \beta_{3}(\text{hp}) + \beta_{4}(\text{drat}) + \beta_{5}(\text{wt}) + \beta_{6}(\text{qsec}) + \beta_{7}(\text{vs}) + \beta_{8}(\text{am}) + \beta_{9}(\text{gear}) + \beta_{10}(\text{carb}) + \epsilon
#> $$

# Works for categorical variables too, putting levels as subscripts
mod3 <- lm(Sepal.Length ~ Sepal.Width + Species, iris)
extract_eq(mod3)

#> $$
#> \text{Sepal.Length} = \alpha + \beta_{1}(\text{Sepal.Width}) + \beta_{2}(\text{Species}_{\text{versicolor}}) + \beta_{3}(\text{Species}_{\text{virginica}}) + \epsilon
#> $$

set.seed(8675309)
d <- data.frame(cat1 = rep(letters[1:3], 100),
                cat2 = rep(LETTERS[1:3], each = 100),
                cont1 = rnorm(300, 100, 1),
                cont2 = rnorm(300, 50, 5),
                out   = rnorm(300, 10, 0.5))
mod4 <- lm(out ~ ., d)
extract_eq(mod4)

#> $$
#> \text{out} = \alpha + \beta_{1}(\text{cat1}_{\text{b}}) + \beta_{2}(\text{cat1}_{\text{c}}) + \beta_{3}(\text{cat2}_{\text{B}}) + \beta_{4}(\text{cat2}_{\text{C}}) + \beta_{5}(\text{cont1}) + \beta_{6}(\text{cont2}) + \epsilon
#> $$

# Don't italicize terms
extract_eq(mod1, ital_vars = FALSE)

#> $$
#> \text{mpg} = \alpha + \beta_{1}(\text{cyl}) + \beta_{2}(\text{disp}) + \epsilon
#> $$

# Wrap equations in an "aligned" environment
extract_eq(mod2, wrap = TRUE)

#> $$
#> \begin{aligned}
#> \text{mpg} &= \alpha + \beta_{1}(\text{cyl}) + \beta_{2}(\text{disp}) + \beta_{3}(\text{hp})\ + \\
#> &\quad \beta_{4}(\text{drat}) + \beta_{5}(\text{wt}) + \beta_{6}(\text{qsec}) + \beta_{7}(\text{vs})\ + \\
#> &\quad \beta_{8}(\text{am}) + \beta_{9}(\text{gear}) + \beta_{10}(\text{carb}) + \epsilon
#> \end{aligned}
#> $$

# Wider equation wrapping
extract_eq(mod2, wrap = TRUE, terms_per_line = 4)

#> $$
#> \begin{aligned}
#> \text{mpg} &= \alpha + \beta_{1}(\text{cyl}) + \beta_{2}(\text{disp}) + \beta_{3}(\text{hp})\ + \\
#> &\quad \beta_{4}(\text{drat}) + \beta_{5}(\text{wt}) + \beta_{6}(\text{qsec}) + \beta_{7}(\text{vs})\ + \\
#> &\quad \beta_{8}(\text{am}) + \beta_{9}(\text{gear}) + \beta_{10}(\text{carb}) + \epsilon
#> \end{aligned}
#> $$

# Include model estimates instead of Greek letters
extract_eq(mod2, wrap = TRUE, terms_per_line = 2, use_coefs = TRUE)

#> $$
#> \begin{aligned}
#> \text{mpg} &= 12.3 - 0.11(\text{cyl})\ + \\
#> &\quad 0.01(\text{disp}) - 0.02(\text{hp})\ + \\
#> &\quad 0.79(\text{drat}) - 3.72(\text{wt})\ + \\
#> &\quad 0.82(\text{qsec}) + 0.32(\text{vs})\ + \\
#> &\quad 2.52(\text{am}) + 0.66(\text{gear})\ - \\
#> &\quad 0.2(\text{carb}) + \epsilon
#> \end{aligned}
#> $$

# Don't fix doubled-up "+ -" signs
extract_eq(mod2, wrap = TRUE, terms_per_line = 4, use_coefs = TRUE, fix_signs = FALSE)

#> $$
#> \begin{aligned}
#> \text{mpg} &= 12.3 + -0.11(\text{cyl}) + 0.01(\text{disp}) + -0.02(\text{hp})\ + \\
#> &\quad 0.79(\text{drat}) + -3.72(\text{wt}) + 0.82(\text{qsec}) + 0.32(\text{vs})\ + \\
#> &\quad 2.52(\text{am}) + 0.66(\text{gear}) + -0.2(\text{carb}) + \epsilon
#> \end{aligned}
#> $$

# Use other model types, like glm
set.seed(8675309)
d <- data.frame(out = sample(0:1, 100, replace = TRUE),
                cat1 = rep(letters[1:3], 100),
                cat2 = rep(LETTERS[1:3], each = 100),
                cont1 = rnorm(300, 100, 1),
                cont2 = rnorm(300, 50, 5))
mod5 <- glm(out ~ ., data = d, family = binomial(link = "logit"))
extract_eq(mod5, wrap = TRUE)

#> $$
#> \begin{aligned}
#> \log\left[ \frac { P( \text{out} = \text{1} ) }{ 1 - P( \text{out} = \text{1} ) } \right] &= \alpha + \beta_{1}(\text{cat1}_{\text{b}}) + \beta_{2}(\text{cat1}_{\text{c}}) + \beta_{3}(\text{cat2}_{\text{B}})\ + \\
#> &\quad \beta_{4}(\text{cat2}_{\text{C}}) + \beta_{5}(\text{cont1}) + \beta_{6}(\text{cont2}) + \epsilon
#> \end{aligned}
#> $$