Display equations in shiny apps

These are a set of functions designed to help render equations in shiny applications (see the vignette about Shiny).

renderEq(expr, env = parent.frame(), quoted = FALSE, outputArgs = list())

eqOutput(outputId)

Arguments

expr

An R expression, specifically a call to extract_eq()

env

The environment

quoted

Is the expresion quoted?

outputArgs

list of output arguments

outputId

The identifier of the output from the server. Should be passed as a string.

Value

Render the equation in a suitable way for Shiny for renderEq() in an eqOutput() equation output element that can be included in a panel.

Functions

renderEq(): Rendering function

eqOutput(): Output function